+++ src/GamIO/ML5Reals.cc
@@ -138,10 +138,10 @@ MatLab5Re::~MatLab5Re() { }
  
 int MatLab5Re::write(std::fstream& fp, const matrix& mx) const	
   {
-  long TB = 9;					// MATLAB data type miDOUBLE
-  fp.write((char*)&TB, sizeof(long));			// Output the data type  (TAG)
+  int32_t TB = 9;					// MATLAB data type miDOUBLE
+  fp.write((char*)&TB, sizeof(int32_t));			// Output the data type  (TAG)
   TB = mx.rows()*mx.cols()*sizeof(double);	// Number of mx bytes to output
-  fp.write((char*)&TB, sizeof(long));			// Output the # of bytes (TAG)
+  fp.write((char*)&TB, sizeof(int32_t));			// Output the # of bytes (TAG)
   int i,j;
   double d;
   for(j=0; j<mx.cols(); j++)			// Output the real elements of
@@ -155,10 +155,10 @@ int MatLab5Re::write(std::fstream& fp, const matrix& mx) const
   
 int MatLab5Re::write(std::fstream& fp, const row_vector& rv) const                      
   {
-  long TB = 9;                                  // MATLAB data type miDOUBLE
-  fp.write((char*)&TB, sizeof(long));                  // Output the data type  (TAG)
+  int32_t TB = 9;                                  // MATLAB data type miDOUBLE
+  fp.write((char*)&TB, sizeof(int32_t));                  // Output the data type  (TAG)
   TB = rv.size()*sizeof(double);                // Number of rv bytes to output
-  fp.write((char*)&TB, sizeof(long));                  // Output the # of bytes (TAG)
+  fp.write((char*)&TB, sizeof(int32_t));                  // Output the # of bytes (TAG)
   double d;
   for(int i=0; i<rv.size(); i++)		// Output the vector elements
     {
@@ -171,10 +171,10 @@ int MatLab5Re::write(std::fstream& fp, const row_vector& rv) const
   
 int MatLab5Re::write(std::fstream& fp, const col_vector& cv) const           
   {
-  long TB = 9;                                  // MATLAB data type miDOUBLE
-  fp.write((char*)&TB, sizeof(long));                  // Output the data type  (TAG)
+  int32_t TB = 9;                                  // MATLAB data type miDOUBLE
+  fp.write((char*)&TB, sizeof(int32_t));                  // Output the data type  (TAG)
   TB = cv.size()*sizeof(double);                // Number of cv bytes to output
-  fp.write((char*)&TB, sizeof(long));                  // Output the # of bytes (TAG)
+  fp.write((char*)&TB, sizeof(int32_t));                  // Output the # of bytes (TAG)
   double d;
   for(int i=0; i<cv.size(); i++) 		// Output the vector elements
     {
